Use images for consistency in the comparison table
I have wrestled much with GitHub’s Markdown formatting over this, and
I’m not altogether pleased with the result: it’s mostly OK, but it
turns the image references into links, which is undesirable. Eventually
I’d like them all to be links to the docs, but unless you have made it a
link yourself, GitHub insists on making it a link.

I decided to go all HTML on it, so that the left hand column can be made
<th>. I adjusted alignment of all the cells for optimality, too.

I omitted the “(see note below)” in the alt text of the partial.svg
cases to minimise necessary reflow when the images load. I don’t want to
specify the intrinsic dimensions on the <img> tag, because some browsers
will constrain the alt text to that size if they fail to load the
images. (Oh to have the SVG *inline*, but no, GFM doesn’t like that.)

Fixes #540.