Stop slugifying the page slug a second time
It’s already a slug because the user wrote it so—or if it doesn’t
conform to what we expect of slugs, the user did so *deliberately* by
naming the file thus. e.g. I have an article on my website with slug
`rust-cfg_attr`, and I definitely don’t want it turned into
`rust-cfg-attr`.

So there’s no need, or desire, to slugify it again.

You may notice I also dropped the `.trim()` here; that’s also
deliberate, because if you were mad enough to put whitespace at the
start or end of the slug, you must have meant it.

This restores consistency with sections, which were already not
slugified. Taxonomy terms are still slugified, for better or for worse
(worse, in my strong opinion, but I’ll address that another time).

Fixes #768.