New config option `author`
This is intended for use in <meta name=author> on normal pages and in
feeds.

I’ve structured things the way I have so it’s easy to use the general
structure of it elsewhere, if you’re overriding templates.

Specifically, I have in mind page-specific authors. I first thought to
add a same-shaped `author` field on pages, but there’s already precedent
for using taxonomies for that, and that’s more powerful, so I recommend
using taxonomies with a straightforward format-shift after this fashion:

Put something like this in config.toml:

```toml
[extra.taxonomies.authors.chris-morgan]
name = "Chris Morgan"
uri = "https://chrismorgan.info"
email = "me@chrismorgan.info"
```

And in the feed template, the likes of this, within the `<entry>` tag:

```tera
{%- for author in page.taxonomies.authors %}
    {{ macros::author(author=config.extra.taxonomies.authors[author]) }}
{%- endfor %}
```

The documentation could do with a recipes section for this and similar.
Templates and config/frontmatter design can get quite complex.

One key thing I’m undecided about in the design of this is allowing a
string at all. Requiring the object might be more sensible, allowing
templates to support it more robustly. But then, I’m not sure whether
`config.author.name` works properly in Tera in the total absence of
author, anyway. Might require more care to make it work ideally.